优化snmp相关函数结构
